agent2/requirements.txt
tensorflow==2.11.0
boto3==1.26.107
Pillow==9.5.0
numpy==1.21.6
pymysql